+++ Task_2_1_1/src/test/java/ru/nsu/cjvth/primes/RunnersTest.java
@@ -8,9 +8,10 @@ import org.junit.jupiter.api.Test;
 
 class RunnersTest {
 
-    private static final List<Integer> input =
-        List.of(5671, 74797, 17449, 67431, 56419, 60961, 69623, 60887, 55713, 7993, 8581, 9875,
-            25867, 28933, 64507, 33923);
+    private static final List<Long> input =
+        List.of(5671L, 74797L, 17449L, 67431L, 56419L, 60961L, 69623L, 60887L, 55713L, 7993L, 8581L,
+            9875L, 25867L, 28933L, 64507L, 33923L);
+
     private static final List<Boolean> expected =
         List.of(false, true, true, false, false, true, true, true, false, true, true, false, true,
             true, false, true);
